# responses/autogen/server/interface.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package server

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Добавление отклика в БД
	// (POST /api/responses/addResponse)
	PostApiResponsesAddResponse(w http.ResponseWriter, r *http.Request)
	// Изменение отклика
	// (PUT /api/responses/edit/{response_id})
	PutApiResponsesEditResponseId(w http.ResponseWriter, r *http.Request, responseId int)
	// Получение ответов по id
	// (GET /api/responses/getResponseByResponseID/{response_id})
	GetApiResponsesGetResponseByResponseIDResponseId(w http.ResponseWriter, r *http.Request, responseId int)
	// Получение ответов по employers id
	// (GET /api/responses/getResponsesByEmployersID/{id})
	GetApiResponsesGetResponsesByEmployersIDId(w http.ResponseWriter, r *http.Request, id int)
	// Получение ответов по username
	// (GET /api/responses/getResponsesByUsername/{username})
	GetApiResponsesGetResponsesByUsernameUsername(w http.ResponseWriter, r *http.Request, username string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Добавление отклика в БД
// (POST /api/responses/addResponse)
func (_ Unimplemented) PostApiResponsesAddResponse(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Изменение отклика
// (PUT /api/responses/edit/{response_id})
func (_ Unimplemented) PutApiResponsesEditResponseId(w http.ResponseWriter, r *http.Request, responseId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получение ответов по id
// (GET /api/responses/getResponseByResponseID/{response_id})
func (_ Unimplemented) GetApiResponsesGetResponseByResponseIDResponseId(w http.ResponseWriter, r *http.Request, responseId int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получение ответов по employers id
// (GET /api/responses/getResponsesByEmployersID/{id})
func (_ Unimplemented) GetApiResponsesGetResponsesByEmployersIDId(w http.ResponseWriter, r *http.Request, id int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Получение ответов по username
// (GET /api/responses/getResponsesByUsername/{username})
func (_ Unimplemented) GetApiResponsesGetResponsesByUsernameUsername(w http.ResponseWriter, r *http.Request, username string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PostApiResponsesAddResponse operation middleware
func (siw *ServerInterfaceWrapper) PostApiResponsesAddResponse(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostApiResponsesAddResponse(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PutApiResponsesEditResponseId operation middleware
func (siw *ServerInterfaceWrapper) PutApiResponsesEditResponseId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "response_id" -------------
	var responseId int

	err = runtime.BindStyledParameterWithOptions("simple", "response_id", chi.URLParam(r, "response_id"), &responseId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "response_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PutApiResponsesEditResponseId(w, r, responseId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetApiResponsesGetResponseByResponseIDResponseId operation middleware
func (siw *ServerInterfaceWrapper) GetApiResponsesGetResponseByResponseIDResponseId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "response_id" -------------
	var responseId int

	err = runtime.BindStyledParameterWithOptions("simple", "response_id", chi.URLParam(r, "response_id"), &responseId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "response_id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApiResponsesGetResponseByResponseIDResponseId(w, r, responseId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetApiResponsesGetResponsesByEmployersIDId operation middleware
func (siw *ServerInterfaceWrapper) GetApiResponsesGetResponsesByEmployersIDId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApiResponsesGetResponsesByEmployersIDId(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetApiResponsesGetResponsesByUsernameUsername operation middleware
func (siw *ServerInterfaceWrapper) GetApiResponsesGetResponsesByUsernameUsername(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "username" -------------
	var username string

	err = runtime.BindStyledParameterWithOptions("simple", "username", chi.URLParam(r, "username"), &username, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "username", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetApiResponsesGetResponsesByUsernameUsername(w, r, username)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/responses/addResponse", wrapper.PostApiResponsesAddResponse)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/api/responses/edit/{response_id}", wrapper.PutApiResponsesEditResponseId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/responses/getResponseByResponseID/{response_id}", wrapper.GetApiResponsesGetResponseByResponseIDResponseId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/responses/getResponsesByEmployersID/{id}", wrapper.GetApiResponsesGetResponsesByEmployersIDId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/responses/getResponsesByUsername/{username}", wrapper.GetApiResponsesGetResponsesByUsernameUsername)
	})

	return r
}